requirements.txt
numpy==1.19.1
opencv-contrib-python==4.3.0.36
Pillow==7.2.0
